requirements.txt
sphinx
catalyst-sphinx-theme
pre-commit
black
mlflow
omegaconf
hydra-core
librosa
soundfile
flake8
unidecode
frozendict
pytest
inflect
onnxruntime
numpy
editdistance
tqdm

# for tensorflow
tensorflow-io

# for pytorch
torch-stft
pytorch-lightning